Migrate code from PV1 Neo Base to core services
